+++ frontend/src/api/assignments.ts
@@ -184,3 +184,45 @@ export const getAssignmentRubric = async (
 
   return await result.json();
 };
+
+export const getAssignmentFirstCommit = async (
+  classroomID: number,
+  assignmentID: number 
+): Promise<Date> => {
+  const response = await fetch(
+    `${base_url}/classrooms/classroom/${classroomID}/assignments/assignment/${assignmentID}/first-commit`,
+    {
+      method: "GET",
+      credentials: "include",
+      headers: {
+        "Content-Type": "application/json",
+      },
+    }
+  );
+  if (!response.ok) {
+    throw new Error("Network response was not ok");
+  }
+  const resp = ((await response.json()) as IAssignmentCommitDate).first_commit_at;
+  return resp;
+};
+
+export const getAssignmentTotalCommits = async (
+  classroomID: number,
+  assignmentID: number 
+): Promise<number> => {
+  const response = await fetch(
+    `${base_url}/classrooms/classroom/${classroomID}/assignments/assignment/${assignmentID}/total-commit`,
+    {
+      method: "GET",
+      credentials: "include",
+      headers: {
+        "Content-Type": "application/json",
+      },
+    }
+  );
+  if (!response.ok) {
+    throw new Error("Network response was not ok");
+  }
+  const resp = await response.json();
+  return resp;
+};
